Move `third_party` to `internal/third_party`
